+++ SIL.DblBundle.Tests/Usx/UsxNodeTests.cs
@@ -48,17 +48,6 @@ namespace SIL.DblBundle.Tests.Usx
 	[TestFixture]
 	public class UsxParaTests
 	{
-		private XmlNodeList m_nodes;
-
-		/// <summary>
-		/// Test fixture setup
-		/// </summary>
-		[OneTimeSetUp]
-		public void TestFixtureSetUp()
-		{
-			m_nodes = UsxDocumentTests.GetChaptersAndParasForMarkOneContaining2Verses();
-		}
-
 		/// <summary>
 		/// Tests that the <see cref="UsxPara"/> constructor throws an <see>ArgumentException</see>
 		/// if the underlying USX node is not a para element.
@@ -66,17 +55,35 @@ namespace SIL.DblBundle.Tests.Usx
 		[Test]
 		public void Constructor_UnderlyingNodeIsNotAParaElement_ThrowsArgumentException()
 		{
-			Assert.That(() => new UsxPara(m_nodes[0]),
+			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
+			xmlDoc.LoadXml("<char style=\"wj\">This is some text</char>");
+			var nodes = xmlDoc.SelectNodes("//char");
+			Assert.That(() => new UsxPara(nodes[0]),
 				Throws.ArgumentException.With.Message.StartsWith("Not a valid para node."));
+		}
+
+		/// <summary>
+		/// Tests that the <see cref="UsxPara"/> constructor throws an
+		/// <see>ArgumentNullException</see> if the underlying node is null.
+		/// </summary>
+		[Test]
+		public void Constructor_NullNode_ThrowsArgumentNullException()
+		{
+			Assert.That(() => new UsxPara(null),
+				Throws.ArgumentNullException);
 		}
 
 		/// <summary>
 		/// Tests that correct StyleTag is obtained from the UsxPara.
 		/// </summary>
-		[Test]
-		public void StyleTag_ChapterStart_GetsCorrectStyleTagFrom()
+		[TestCase("p")]
+		[TestCase("q1")]
+		public void StyleTag_NormalPara_GetsCorrectStyleTagFrom(string style)
 		{
-			Assert.AreEqual("p", new UsxPara(m_nodes[1]).StyleTag);
+			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
+			xmlDoc.LoadXml($"<para style=\"{style}\">This is some text</para>");
+			var usxDoc = new UsxDocument(xmlDoc);
+			Assert.AreEqual(style, new UsxPara(usxDoc.GetChaptersAndParas()[0]).StyleTag);
 		}
 	}
 
@@ -87,8 +94,8 @@ namespace SIL.DblBundle.Tests.Usx
 	public class UsxCharTests
 	{
 		/// <summary>
-		/// Tests that the <see cref="UsxPara"/> constructor throws an <see>ArgumentException</see>
-		/// if the underlying USX node is not a para element.
+		/// Tests that the <see cref="UsxChar"/> constructor throws an <see>ArgumentException</see>
+		/// if the underlying USX node is not a "char" element.
 		/// </summary>
 		[Test]
 		public void Constructor_UnderlyingNodeIsNotACharElement_ThrowsArgumentException()
@@ -101,16 +108,29 @@ namespace SIL.DblBundle.Tests.Usx
 		}
 
 		/// <summary>
-		/// Tests that correct StyleTag is obtained from the UsxPara.
+		/// Tests that the <see cref="UsxChar"/> constructor throws an
+		/// <see>ArgumentNullException</see> if the underlying node is null.
 		/// </summary>
 		[Test]
-		public void StyleTag_ChapterStart_GetsCorrectStyleTagFrom()
+		public void Constructor_NullNode_ThrowsArgumentNullException()
+		{
+			Assert.That(() => new UsxChar(null),
+				Throws.ArgumentNullException);
+		}
+
+		/// <summary>
+		/// Tests that correct StyleTag is obtained from the UsxChar.
+		/// </summary>
+		[TestCase("add")]
+		[TestCase("wj")]
+		public void StyleTag_ChapterStart_GetsCorrectStyleTagFrom(string charStyle)
 		{
 			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
-			xmlDoc.LoadXml("<para style=\"p\">This is some text that should <char style=\"add\">be</char> good.</para>");
+			xmlDoc.LoadXml("<para style=\"p\">This is some text that should " +
+				$"<char style=\"{charStyle}\">be</char> good.</para>");
 			var usxDoc = new UsxDocument(xmlDoc);
 			var para = new UsxPara(usxDoc.GetChaptersAndParas()[0]);
-			Assert.AreEqual("add", new UsxChar(para.ChildNodes[1]).StyleTag);
+			Assert.AreEqual(charStyle, new UsxChar(para.ChildNodes[1]).StyleTag);
 		}
 	}
 
@@ -120,20 +140,9 @@ namespace SIL.DblBundle.Tests.Usx
 	[TestFixture]
 	public class UsxChapterTests
 	{
-		private XmlNodeList m_nodes;
-
 		/// <summary>
-		/// Test fixture setup
-		/// </summary>
-		[OneTimeSetUp]
-		public void TestFixtureSetUp()
-		{
-			m_nodes = UsxDocumentTests.GetChaptersAndParasForMarkOneContaining2Verses(true);
-		}
-
-		/// <summary>
-		/// Tests that Constructor throws an <see>ArgumentException</see> if
-		/// the underlying USX node is not a chapter node.
+		/// Tests that <see cref="UsxChapter"/> constructor throws an <see>ArgumentException</see>
+		/// if the underlying USX node is not a chapter node.
 		/// </summary>
 		[Test]
 		public void Constructor_NotAChapterNode_ThrowsArgumentException()
@@ -144,86 +153,110 @@ namespace SIL.DblBundle.Tests.Usx
 			Assert.That(() => { new UsxChapter(usxDoc.GetChaptersAndParas()[0]); },
 				Throws.ArgumentException);
 		}
-		
+
 		/// <summary>
-		/// Tests that correct StyleTag is obtained from the UsxChapter.
+		/// Tests that the <see cref="UsxChapter"/> constructor throws an
+		/// <see>ArgumentNullException</see> if the underlying node is null.
 		/// </summary>
 		[Test]
-		public void StyleTag_ChapterStart_GetsCorrectStyleTag()
+		public void Constructor_NullNode_ThrowsArgumentNullException()
 		{
-			Assert.AreEqual("c", new UsxChapter(m_nodes[0]).StyleTag);
+			Assert.That(() => new UsxChapter(null),
+				Throws.ArgumentNullException);
+		}
+		
+		/// <summary>
+		/// Tests that <see cref="UsxChapter.StyleTag"/> returns "c" when the object is based on a
+		/// chapter start node.
+		/// </summary>
+		[TestCase()]
+		[TestCase(" sid=\"LUK 2\"")]
+		public void StyleTag_ChapterStart_GetsCorrectStyleTag(string sidAttr = null)
+		{
+			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
+			xmlDoc.LoadXml($"<chapter number=\"2\" style=\"c\"{sidAttr}/>");
+			var nodes = xmlDoc.SelectNodes("//chapter");
+			Assert.AreEqual("c", new UsxChapter(nodes[0]).StyleTag);
 		}
 
 		/// <summary>
-		/// Tests that StyleTag returns null if the UsxNode is a chapter end.
+		/// Tests that <see cref="UsxChapter.StyleTag"/> returns null when the object is based on a
+		/// chapter end node.
 		/// </summary>
 		[Test]
 		public void StyleTag_ChapterEnd_ReturnsNull()
 		{
-			var usxNode = new UsxChapter(m_nodes[m_nodes.Count - 1]);
-			Assert.IsNull(usxNode.StyleTag);
+			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
+			xmlDoc.LoadXml("<chapter eid=\"REV 6\"/>");
+			var nodes = xmlDoc.SelectNodes("//chapter");
+			Assert.IsNull(new UsxChapter(nodes[0]).StyleTag);
 		}
 
 		/// <summary>
-		/// Tests that IsChapterStart returns <c>true</c> when the object is based on a node that
-		/// does not have an sid or eid attribute.
+		/// Tests that <see cref="UsxChapter.IsChapterStart"/> returns <c>true</c> when the object
+		/// is based on a chapter start node.
 		/// </summary>
-		[Test]
-		public void IsChapterStart_StartNoSid_ReturnsTrue()
-		{
-			var usxChapterNode = new UsxChapter(m_nodes[0]);
-			Assert.IsTrue(usxChapterNode.IsChapterStart);
-		}
-
-		/// <summary>
-		/// Tests that IsChapterStart returns <c>true</c> when the object is based on a node that
-		/// has an sid attribute.
-		/// </summary>
-		[Test]
-		public void IsChapterStart_StartWithSid_ReturnsTrue()
+		[TestCase()]
+		[TestCase(" sid=\"LUK 2\"")]
+		public void IsChapterStart_ChapterStart_ReturnsTrue(string sidAttr = null)
 		{
 			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
-			xmlDoc.LoadXml("<chapter number=\"2\" style=\"c\" sid=\"MRK 2\"/>");
-			var usxDoc = new UsxDocument(xmlDoc);
-			var usxChapterNode = new UsxChapter(usxDoc.GetChaptersAndParas()[0]);
+			xmlDoc.LoadXml($"<chapter number=\"2\" style=\"c\"{sidAttr}/>");
+			var nodes = xmlDoc.SelectNodes("//chapter");
+			var usxChapterNode = new UsxChapter(nodes[0]);
 			Assert.IsTrue(usxChapterNode.IsChapterStart);
 		}
 
 		/// <summary>
-		/// Tests that IsChapterStart returns <c>true</c> when the object is based on a node that
-		/// has an sid attribute.
+		/// Tests that <see cref="UsxChapter.IsChapterStart"/> returns <c>false</c> when the object
+		/// is based on a chapter end node.
 		/// </summary>
 		[Test]
 		public void IsChapterStart_End_ReturnsFalse()
 		{
-			var usxChapter = new UsxChapter(m_nodes[m_nodes.Count - 1]);
+			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
+			xmlDoc.LoadXml("<chapter eid=\"ACT 27\"/>");
+			var nodes = xmlDoc.SelectNodes("//chapter");
+			var usxChapter = new UsxChapter(nodes[0]);
 			Assert.IsFalse(usxChapter.IsChapterStart);
 		}
 
 		/// <summary>
-		/// Texts that the correct chapter number is obtained from the UsxChapter.
+		/// Tests that <see cref="UsxChapter.ChapterNumber"/> gets the chapter number from the
+		/// number attribute if the node is a chapter start.
 		/// </summary>
-		[Test]
-		public void ChapterNumber_StartNoSid_GetsCorrectChapterNumberFromNumberAttribute()
+		[TestCase()]
+		[TestCase(" sid=\"LUK 2\"")] // Note: This is bad data
+		[TestCase(" sid=\"LUK 3\"")]
+		public void ChapterNumber_ChapterStart_GetsCorrectChapterNumberFromNumberAttribute(
+			string sidAttr = null)
 		{
-			var usxChapter = new UsxChapter(m_nodes[0]);
-			Assert.AreEqual("1", usxChapter.ChapterNumber);
+			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
+			xmlDoc.LoadXml($"<chapter number=\"3\" style=\"c\"{sidAttr}/>");
+			var nodes = xmlDoc.SelectNodes("//chapter");
+			var usxChapter = new UsxChapter(nodes[0]);
+			Assert.AreEqual("3", usxChapter.ChapterNumber);
 		}
 
 		/// <summary>
-		/// Tests that StyleTag returns null if the UsxNode is a chapter end.
+		/// Tests that <see cref="UsxChapter.ChapterNumber"/> gets the chapter number from the eid
+		/// attribute if the node is a chapter end.
 		/// </summary>
-		[Test]
-		public void ChapterNumber_ChapterEnd_GetsCorrectChapterNumberFromEidAttribute()
+		[TestCase("1")]
+		[TestCase("13")]
+		public void ChapterNumber_ChapterEnd_GetsCorrectChapterNumberFromEidAttribute(string chapter)
 		{
-			var usxChapter = new UsxChapter(m_nodes[m_nodes.Count - 1]);
-			Assert.AreEqual("1", usxChapter.ChapterNumber);
+			var xmlDoc = new XmlDocument { PreserveWhitespace = true };
+			xmlDoc.LoadXml($"<chapter eid=\"ACT {chapter}\"/>");
+			var nodes = xmlDoc.SelectNodes("//chapter");
+			var usxChapter = new UsxChapter(nodes[0]);
+			Assert.AreEqual(chapter, usxChapter.ChapterNumber);
 		}
 
 		/// <summary>
-		/// Tests that IsChapterStart throws a <see>NullReferenceException</see> if
-		/// the underlying USX node is a chapter with neither a "number" attribute
-		/// nor an "eid" attribute.
+		/// Tests that <see cref="UsxChapter.ChapterNumber"/> throws a
+		/// <see>NullReferenceException</see> if the underlying USX node is a chapter with neither
+		/// a "number" attribute nor an "eid" attribute.
 		/// </summary>
 		[Test]
 		public void ChapterNumber_InvalidUnderlyingNode_ThrowsNullReferenceException()